backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Data processing
pandas==2.1.3
numpy==1.25.2
pillow==10.1.0
librosa==0.10.1
scipy==1.11.4

# Speech analysis specific
praat-parselmouth==0.4.3
soundfile==0.12.1
audioread==3.0.0
resampy==0.4.2

# Machine Learning
tensorflow==2.15.0
scikit-learn==1.3.2
torch==2.1.1
torchvision==0.16.1

# HTTP client
httpx==0.25.2
aiofiles==23.2.1

# Cloud storage (AWS S3)
boto3==1.34.0
botocore==1.34.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Monitoring
prometheus-client==0.19.0

# OpenAI for RAG
openai==1.3.6